requirements.txt
Flask
Flask-Caching
urllib.parse
redis
python-dotenv
pandas
numpy
scikit-learn
joblib
gunicorn